Adding tests for relay service. Fixing small issues.
