+++ package.json
@@ -53,6 +53,7 @@
     "jest": "^26.6.3",
     "near-web3-provider": "^1.0.6",
     "rlp": "^2.2.6",
+    "supertest": "^6.0.1",
     "ts-jest": "^26.4.4",
     "yargs": "^16.2.0"
   }
